fix: ensure `connectedCallback` runs after Ember initializer
This resolves an issue where the initialization logic was being run in
the constructor of the custom element _before_ the Ember initializer was
finished. The result of this issue was that the custom element attempted
to access properties that were not yet ready for access.

The Ember initializer creates a new promise on each run to ensure that,
in a testing scenario, each initializer run gets it's own promise. This
isn't strictly necessary, but is probably safer for testing purposes.

Closes #6